bgp: add fine-grained debug message control
Enable users to select which log messages they want to see, instead
of logging everything by default.

Debugging of sent and received packets can be configured at both
the instance and neighbor levels, with neighbor-level configuration
taking precedence. The YANG modeling of the debug settings, using
the trace-options container with a list of flags, is inspired by the
Juniper and SR Linux YANG models.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>