bgp, vrrp: honor config option for logging state changes
The BGP YANG module includes a log-neighbor-state-changes leaf that
controls whether peer state changes should be logged. While we already
support fine-grained logging via trace-options, we must also honor
this config knob, which is enabled by default. The same applies to
the VRRP module.

Ideally, the log level would be INFO when log-neighbor-state-changes is
enabled, and DEBUG otherwise, when only trace-options are in use. For
now, we log all state changes at the DEBUG level.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>