*: rework support for blackhole routes
In the "ietf-routing" YANG module, a blackhole route is modeled as a
normal route with a special kind of nexthop (the "special-next-hop"
leaf), which precludes the existence of other nexthops.

In our codebase, the `Nexthop` enum had a `Special` variant that was
used to represent such special nexthop types (blackhole, unreachable,
and prohibit). While this maps nicely to the YANG model, it allows the
possibility of ECMP where one of the nexthops is a blackhole nexthop,
which would be invalid.

In this commit, we remove the `Special` variant from the `Nexthop`
enum.  Instead, each route now has a separate field indicating the
type of the route (normal unicast, blackhole, etc.). This maps better
to reality, where blackhole is a property of a route, not of a nexthop.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>