isis: convert packet parser to use fallible `try_get_` methods
Update the packet parser to use the new fallible `try_get_` methods
instead of the infallible ones that will panic when attempting to
read out of bounds.

While the parser already includes checks to prevent out-of-bounds
reads, using the fallible methods adds an extra layer of safety.

As with other parsing errors, malformed TLVs are ignored, and
malformed PDUs are rejected.

This change does not affect performance. Both `get_` and `try_get_`
perform bounds checking, but `get_` panics on failure, whereas
`try_get_` returns an error if there are not enough bytes remaining
to read.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>