isis: support neighbor reachability detection via BFD
Add support for using BFD to detect neighbor reachability. For
dual-stack adjacencies, separate BFD sessions are maintained for IPv4
and IPv6.

BFD can be enabled on a per-interface basis, with separate options
to configure BFD-specific parameters such as intervals and local
multiplier.

In single-topology mode, the adjacency is considered up only if all
BFD sessions (IPv4 and IPv6) are up.

In multi-topology mode (not yet implemented), the adjacency is torn
down only if all BFD sessions are down. BFD session status will also
be checked when adding IS reachability information to the local LSP
during origination.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>