isis: add support for multi-topology routing
This commit implements support for multi-topology routing as specified
by RFC 5120.

The MT TLVs were implemented in a previous commit. This commit adds:
* Processing of MT configuration options
* Tracking of MT-IDs announced by each adjacency
* Addition of MT-IDs in Hello PDUs
* Addition of MT TLVs to the local LSP
* Separate SPF runs per topology, using the appropriate TLVs
* Independent handling of overload and attach bits for each topology
* Propagation of MT reachability TLVs from L1 to L2 on L1/L2 routers

Currently, only the IPv6 unicast topology can be configured, in
addition to the standard topology that is enabled by default.

Implementing MT adds some unwelcome complexity to the code. A few
things could have been done in a better way, but we can do the
necessary cleanups over time.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>