isis: add MT test topology
This is a variation of the second topology, using MT to configure
non-congruent IPv4 and IPv6 routing.

In this scenario, all routers participate in the IPv4 topology,
while only selected routers (rt1, rt3, rt5, and rt6) participate
in the IPv6 topology.

As expected:
- rt1 reaches rt6 via IPv4 using ECMP through rt2 and rt3.
- rt1 reaches rt6 via IPv6 through rt3 only.
- rt5 (an L1/L2 router) sets the ATT bit for the IPv6 topology.
- rt5 propagates MT IPv6 reachability TLVs from L1 to L2.
- rt6 computes a default IPv4 route using ECMP through rt4 and rt5.
- rt6 computes a default IPv6 route through rt5 only.

MT-specific conformance tests will be added later to further validate
MT behavior.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>