ospf: Add RFC 5613 support (Link Local Signaling)
This commit introduces LLS data block appended to OSPF Hello and DbDesc
packets. The current code is shared between OSPFv2 and OSPFv3.

- About authentication:
For OSPFv{2,3}, if the authentication is disabled, the LLS
data block checksum is verified.

For OSPFv2, when the authentication is enabled, the CA-TLV digest is
verified.

For OSPFv3, when the authentication is enabled, there is nothing to do
as the LLS data block is included in the authentication digest. This
behavior is documented in RFC 7166 Section 2.

- About enabling LLS data blocks through YANG:
LLS data blocks are enabled per interface. That enables generating LLS
data block for Hello and DbDesc messages orginated for a that specific
interface. If some LLS data is specified through OSPF configuration, the
L-bit is enabled in the packet options.

This feature also enables examining LLS data block for received messages
if the option is enabled and the L-bit flag is present in the message
options.

Currently, the operation detailed above are mainly placeholders for future
behavior since RFC 5613 does not mention how to configure the TLVs it
specifies.

- About the presence of EO TLV in LlsDbDescData:
RFC 4811 Section 2.1 specifies that EO TLV may be included in DbDesc
packets.

Signed-off-by: Nicolas Rybowski <nicolas.rybowski@uclouvain.be>
---
v0 -> v1: Fixed nits mentioned in Renato's review at
  https://github.com/holo-routing/holo/pull/77