ospf: Add initial RFC 5613 support
This commit introduces LLS data block appended to
OSPF Hello and DbDesc packets. The current code is
shared between OSPFv2 and OSPFv3.

Currently, the LLS data block checksum is not
computed yet, the block is not authentified if
authentication is enabled (OSPFv2 only), and
CA-TLV is not implemented (OSPFv2 only).

Signed-off-by: Nicolas Rybowski <nicolas.rybowski@uclouvain.be>