ospf: Enable LLS YANG feature
This feature allows configuring LLS per interace.

That enables generating LLS data block for Hello and DbDesc messages when
the option is enabled on the interface. In that case, if some LLS data
is specified, the L-bit is enabled in the packet options.

This feature also enables examining LLS data block for received messages
if the option is enabled and the L-bit flag is present in the message
options.

Currently, the operation detailed above are mainly placeholders for future
behavior since RFC 5613 does not mention how to configure the TLVs it
specifies.

Signed-off-by: Nicolas Rybowski <nicolas.rybowski@uclouvain.be>