daemon: remove chroot jail
The chroot mechanism was designed for filesystem isolation, not for
security. If the daemon process were compromised, which is unlikely
thanks to Rust's memory safety, an attacker could still escape the
jail through various means. Remove the chroot jail, as it is overly
restrictive and provides little to no security benefit. Running
with dropped privileges and a very limited set of Linux capabilities
already offers a strong level of security.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>