protocols: rename `southbound` module to `ibus`
Each protocol implementation operates as an event loop that processes
three types of events:
* Northbound events (e.g. config changes, RPCs, state queries)
* Protocol events (e.g. received packets, timers)
* ibus messages (e.g. interface status updates, route redistribution)

In this commit, we rename the `southbound` module of each protocol
crate to `ibus`, as the latter is a more general term. We also
move all ibus message handling into the `ibus` module to improve
organization. As a result, the `events.rs` files now contain only
the handling of protocol events.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>