northbound: refactor `callbacks()` internal APIs
Remove the `Option` wrapper from the return type, requiring
implementers to provide a concrete `Callbacks` value, even if it's
an empty list.

This simplifies some code in the northbound layer and makes future
extensions easier to implement.  For example, holo-interface currently
doesn't implement any RPCs, but when one is needed, the required
boilerplate will already be in place.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>