isis, ospf: fix database clearing RPC
In holo-isis and holo-ospf, LSPs and LSAs are stored in `Lsdb` structs
backed by an arena. Clearing an `Lsdb` isn't as simple as resetting
the struct, we also need to remove the corresponding LSP/LSA entries
from the arena.

Also, when clearing an `Lsdb`, we should be careful to not reset the
`next_id` field to ensure all LSP and LSA entries have unique IDs.

Issue discovered with IxANVL.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>