isis: implement three-way handshake for point-to-point adjacencies
Implement the three-way adjacency handshake for point-to-point links
as specified in RFC 5303. Adjacencies are only established when
bidirectional communication is confirmed, preventing formation over
unidirectional links.

Interop with Junos and Arista cEOS was verified. tc-netem(8) was
used to simulate unidirectional links and confirm that the handshake
prevents adjacency formation until the link is fully bidirectional.

As the three-way handshake mechanism is backwards compatible, it
doesn't need to be configured and is enabled by default.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>