ci: use fixed nightly version for fuzz targets build check
A recent change in nightly caused rustc to emit new warnings in an
upstream crate we have no control over. To fix this issue, and to
prevent similar cases in the future, use a fixed nightly version for
the fuzz targets build check, which is the only job that requires
nightly Rust. This avoids sporadic CI failures that we cannot do
anything about.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>