isis: store parent vertices in the SPT
Flooding reduction algorithms need to check whether one node lies
on the shortest path toward another. To support these path checks,
SPT vertices are now stored in an arena, and each vertex keeps a
list of its parent indices.

Although the SPT is conceptually a tree, IS-IS and OSPF can assign
multiple parents to a vertex to represent equal-cost paths. For
this reason, we cannot use a regular tree data structure to store
the SPT vertices.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>