isis: implement distributed flooding reduction
This is an implementation of the flooding reduction mechanism
specified in draft-ietf-lsr-distoptflood-11. That draft defines a
distributed algorithm that can drastically reduce flooding overhead
in densely meshed networks, such as the Clos fabrics common in
datacenter networks, and can improve convergence times as a result.
In preliminary tests, flooding volume can be reduced by up to 90%
in three-tier Clos topologies.

These optimizations are currently disabled by default and must be
explicitly enabled in the configuration.

While the framework described in
draft-prz-lsr-interop-flood-reduction-architecture-01 allows for
the use of different flooding reduction algorithms, no attempt was
made to create a unified API where new algorithms can be plugged
in easily. The reason is that future algorithms may work in very
different ways, and forcing a generic abstraction around a single
algorithm would not be useful.

This commit also includes a few updates to the SPT computation
function, such as adding support for processing all MT-IDs and
ignoring the overload bit. These changes are not in the current
draft but are expected to be documented in the next version, as
discussed on the LSR mailing list.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>