northbound: refactor and optimize configuration validation
In addition to YANG validation, Holo uses code-level validation in
the form of callbacks to ensure a given configuration is valid. This
is necessary to express more complex validation constraints, or
to reject configuration values that are not currently supported.

In the current implementation, configuration validation is
performed in parallel, with each component validating its associated
configuration subtree. However, this has a few problems. First,
validation callbacks aren't invoked when the corresponding component
hasn't started yet, such as OSPF validation callbacks for an OSPF
instance that isn't configured yet. Second, validation callbacks
do not depend on existing state, making it simpler to validate the
entire configuration in a single place.

This commit refactors the code so that holod performs full
configuration validation itself, instead of sending validation
requests to internal components. This approach is simpler and fixes
the aforementioned bug.

This commit also introduces an optimization where, instead of
traversing all configuration nodes and validating them one by one,
the list of validation callbacks is iterated over and only the
corresponding nodes in the configuration are validated. This scales
much better for larger configurations.

Fixes: #97
Fixes: #98

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>